requirements.txt
Faker==37.11.0
playwright==1.55.0
pytest==8.4.2
pytest-playwright==0.7.1
python-dotenv==1.1.1
